Clone detection refactor
