+++ src/main/java/org/jetbrains/research/anticopypaster/cloneprocessors/Clone.java
@@ -1,0 +1,8 @@
+package org.jetbrains.research.anticopypaster.cloneprocessors;
+
+import com.intellij.psi.PsiElement;
+
+import java.util.List;
+
+public record Clone(PsiElement start, PsiElement end, List<Variable> liveVars) {}
+
