+++ src/main/java/org/jetbrains/research/anticopypaster/cloneprocessors/Variable.java
@@ -1,0 +1,3 @@
+package org.jetbrains.research.anticopypaster.cloneprocessors;
+
+public record Variable(String identifier, String type) {}
